'commit_5.22.19'

+++ tracker/urlcheck.py
@@ -1,8 +1,5 @@
 from .models import InstaPost, PostRecord
-from PIL import Image
-from io import BytesIO
-import requests
-import threading
+import requests, threading, fleep
 
 
 # Utilizes multithreading to check for dead url links and delete corresponding objects
@@ -16,21 +13,28 @@ def dead_url_check():
             post = url_list.pop()
             # Tries to open each Instagram post picture and thumbnail picture
             try:
-                cont = requests.get(post.pic_url).content
-                Image.open(BytesIO(cont))
-                cont = requests.get(post.thumb_url).content
-                Image.open(BytesIO(cont))
-                print(threading.current_thread().name + ': both pic and thumb URLs live for post in ' + post.location_text)
-
-            # If the Instagram picture link is dead, deletes post so that it can be refreshed
+                # Main image
+                main_cont = requests.get(post.pic_url).content
+                main_kind = fleep.get(main_cont)
+                # Thumbnail image
+                thumb_cont = requests.get(post.thumb_url).content
+                thumb_kind = fleep.get(thumb_cont)
+                # If the URL leads to a media file, do nothing
+                if main_kind.type and thumb_kind.type:
+                    print(threading.current_thread().name + ': both ' + str(main_kind.type) + ' and ' +
+                          str(thumb_kind.type) + ' URLs live for post in ' + post.location_text)
+                # If the URL does not lead to a media file, deletes posts so that it can be refreshed by instacrape.py
+                else:
+                    print(threading.current_thread().name + ': dead url, deleting post and post record for id: ' + str(
+                        post.post_id))
+                    post.delete()
+                    try:
+                        rec = PostRecord.objects.get(post_id=post.post_id)
+                        rec.delete()
+                    except:
+                        print("error deleting post record.")
             except:
-                print(threading.current_thread().name + ': dead url, deleting post and post record for id: ' + str(post.post_id))
-                post.delete()
-                try:
-                    rec = PostRecord.objects.get(post_id=post.post_id)
-                    rec.delete()
-                except:
-                    print("error deleting post record.")
+                print('error')
 
     t1 = threading.Thread(target=thread, name='thread 1')
     t2 = threading.Thread(target=thread, name='thread 2')
